+++ src/Bio/Graph/ReferenceDAG/Internal.hs
@@ -10,6 +10,9 @@
 --
 -----------------------------------------------------------------------------
 
+{-# OPTIONS_GHC -Wno-unused-matches #-}
+{-# OPTIONS_GHC -Wno-name-shadowing #-}
+
 {-# LANGUAGE DeriveFunctor              #-}
 {-# LANGUAGE DeriveGeneric              #-}
 {-# LANGUAGE FlexibleContexts           #-}
@@ -23,20 +26,21 @@
 
 module Bio.Graph.ReferenceDAG.Internal where
 
+import           Analysis.Parsimony.Internal
 import           Bio.Graph.BinaryRenderingTree
 import           Bio.Graph.Component
 import           Bio.Graph.LeafSet
+import           Control.Applicative           as Alt (Alternative (empty, (<|>)))
 import           Control.Arrow                 ((&&&), (***))
 import           Control.DeepSeq
 import           Control.Lens                  as Lens (Lens, Lens', lens, to)
 import           Control.Lens.Fold             (Fold, folding)
-import           Control.Lens.Operators        ((%~), (.~))
+import           Control.Lens.Operators        ((%~), (.~), (^.))
 import           Control.Monad.State.Lazy
 import           Data.Bifunctor
 import           Data.EdgeSet
 import           Data.Foldable
 import           Data.Foldable.Custom
-import           Data.Functor                  ((<$))
 import           Data.GraphViz.Attributes
 import           Data.GraphViz.Printing
 import           Data.GraphViz.Types           hiding (attrs)
@@ -48,7 +52,7 @@ import qualified Data.IntMap                   as IM
 import           Data.IntSet                   (IntSet)
 import qualified Data.IntSet                   as IS
 import           Data.Key
-import           Data.List                     (intercalate)
+import           Data.List                     (intercalate, tails)
 import           Data.List.NonEmpty            (NonEmpty (..), intersperse)
 import qualified Data.List.NonEmpty            as NE
 import           Data.List.Utility             (isSingleton)
@@ -62,10 +66,12 @@ import           Data.String
 import           Data.Traversable
 import           Data.Tree                     (unfoldTree)
 import           Data.Tree.Pretty              (drawVerticalTree)
+import           Data.Tuple.Utility
 import           Data.Vector                   (Vector)
 import qualified Data.Vector                   as V
 import qualified Data.Vector.Custom            as V (fromList')
 import           Data.Vector.Instances         ()
+import           Data.Vector.Utility           as DV
 import           GHC.Generics
 import           Numeric.Extended.Real
 import           Prelude                       hiding (lookup, zipWith)
@@ -76,7 +82,7 @@ import           Text.XML.Custom
 -- |
 -- A constant time access representation of a directed acyclic graph.
 data  ReferenceDAG d e n
-    = RefDAG
+    = ReferenceDAG
     { references :: {-# UNPACK #-} !(Vector (IndexData e n))
     , rootRefs   :: !(NonEmpty Int)
     , graphData  :: GraphData d
@@ -277,7 +283,7 @@ type instance Key (ReferenceDAG d e) = Int
 instance Bifunctor (ReferenceDAG d) where
 
     bimap f g dag =
-        RefDAG
+        ReferenceDAG
         { references = h <$> references dag
         , rootRefs   = rootRefs  dag
         , graphData  = graphData dag
@@ -306,7 +312,7 @@ instance FoldableWithKey (ReferenceDAG d e) where
 instance Functor (ReferenceDAG d e) where
 
     fmap f dag =
-        RefDAG
+        ReferenceDAG
         { references = g <$> references dag
         , rootRefs   = rootRefs  dag
         , graphData  = graphData dag
@@ -321,7 +327,7 @@ instance HasLeafSet (ReferenceDAG d e n) (LeafSet n) where
     leafSet = Lens.to getter
         where
             getter :: ReferenceDAG d e n -> LeafSet n
-            getter (RefDAG v _ _) = LeafSet $ foldMap f v
+            getter (ReferenceDAG v _ _) = LeafSet $ foldMap f v
 
             f e | null (childRefs e) = [nodeDecoration e]
                 | otherwise          = mempty
@@ -574,7 +580,7 @@ connectEdge dag originTransform targetTransform (ooRef, otRef) (toRef, ttRef) = 
     oNewRef = oldLen -- synonym
     tNewRef = oldLen + 1
     newDag  =
-      RefDAG
+      ReferenceDAG
         <$> const newVec
         <*> rootRefs
         <*> graphData
@@ -629,7 +635,7 @@ invadeEdge dag transformation node (oRef, iRef) = newDag
     newRef  = oldLen -- synonym
     refs    = references dag
     newDag  =
-      RefDAG
+      ReferenceDAG
         <$> const newVec
         <*> rootRefs
         <*> graphData
@@ -692,13 +698,6 @@ contractToContiguousVertexMapping inputMap = foldMapWithKey contractIndices inpu
 defaultGraphMetadata :: Monoid m => GraphData d -> GraphData m
 {-# INLINE defaultGraphMetadata #-}
 defaultGraphMetadata = _graphMetadata .~ mempty
-{--    GraphData
-      <$> dagCost
-      <*> networkEdgeCost
-      <*> rootingCost
-      <*> totalBlockCost
-      <*> const mempty
---}
 
 -- |
 -- Overwrite the current graph metadata with a default value.
@@ -707,11 +706,20 @@ defaultGraphMetadata = _graphMetadata .~ mempty
 defaultMetadata :: Monoid m => ReferenceDAG d e n -> ReferenceDAG m e n
 {-# INLINE defaultMetadata #-}
 defaultMetadata = _graphData %~ defaultGraphMetadata
-{--    RefDAG
-      <$> references
-      <*> rootRefs
-      <*> defaultGraphMetadata . graphData
---}
+
+-- |
+-- Zero cost graph data with 'graphMetadata' set to 'mempty'.
+zeroCostGraphData :: Monoid m => GraphData m
+{-# INLINE zeroCostGraphData #-}
+zeroCostGraphData
+  = GraphData
+  { dagCost         = 0
+  , networkEdgeCost = 0
+  , rootingCost     = 0
+  , totalBlockCost  = 0
+  , graphMetadata   = mempty
+  }
+
 
 -- |
 -- Ensure that each vertex has either:
@@ -982,7 +990,7 @@ nodeFoldMap f = foldMap f . fmap nodeDecoration . references
 -- a list of child node decorations with the logic function already applied,
 -- and returns the new decoration for the current node.
 nodePostOrder :: (n -> [n'] -> n') -> ReferenceDAG d e n -> ReferenceDAG d e n'
-nodePostOrder f dag = RefDAG <$> const newReferences <*> rootRefs <*> graphData $ dag
+nodePostOrder f dag = ReferenceDAG <$> const newReferences <*> rootRefs <*> graphData $ dag
   where
     dagSize       = length $ references dag
     newReferences = V.generate dagSize h
@@ -1004,7 +1012,7 @@ nodePostOrder f dag = RefDAG <$> const newReferences <*> rootRefs <*> graphData 
 -- a list of parent node decorations with the logic function already applied,
 -- and returns the new decoration for the current node.
 nodePreOrder :: (n -> [(Word, n')] -> n') -> ReferenceDAG d e n -> ReferenceDAG d e n'
-nodePreOrder f dag = RefDAG <$> const newReferences <*> rootRefs <*> graphData $ dag
+nodePreOrder f dag = ReferenceDAG <$> const newReferences <*> rootRefs <*> graphData $ dag
   where
     dagSize       = length $ references dag
     newReferences = V.generate dagSize h
@@ -1082,7 +1090,7 @@ topologyRendering dag = drawVerticalTree . unfoldTree f . NE.head $ rootRefs dag
 -- * Does not normalize nodes for proper out-degree values.
 fromList :: Foldable f => f (IntSet, n, IntMap e) -> ReferenceDAG () e n
 fromList xs =
-    RefDAG
+    ReferenceDAG
     { references = referenceVector
     , rootRefs   = rootSet
     , graphData  = GraphData 0 0 0 0 ()
@@ -1113,7 +1121,7 @@ unfoldDAG :: (Eq a, Hashable a, Monoid e, Monoid n)
           -> a                              -- ^ Seed value
           -> ReferenceDAG () e n
 unfoldDAG f origin =
-    RefDAG
+    ReferenceDAG
     { references = referenceVector
     , rootRefs   = NE.fromList roots2 -- otoList rootIndices
     , graphData  = GraphData 0 0 0 0 ()
@@ -1241,131 +1249,6 @@ getDotContext uniqueIdentifierBase mostSignificantDigit dag = second mconcat . u
 
 
 -- |
--- Generate the set of candidate network edges for a given DAG.
-candidateNetworkEdges :: ReferenceDAG d e n -> Set ( (Int, Int), (Int,Int) )
-candidateNetworkEdges dag = S.filter correctnessCriterion $ foldMapWithKey f mergedVector
-  where
-    mergedVector  = zipWith mergeThem ancestoralEdgeSets descendantEdgeSets
-
-    mergeThem a d =
-        IndexData
-        { nodeDecoration = nodeDecoration a
-        , parentRefs     = parentRefs a
-        , childRefs      = zipWith (<>) (childRefs a) (childRefs d)
-        }
-
-    correctnessCriterion x = doesNotShareNode x && notNetworkEdges x
-
-    doesNotShareNode ((a,b),(c,d)) = a /= c && a /= d && b /= c && b /= d
-
-    notNetworkEdges  ((_,b),(_,d)) = isNotNetworkNode b && isNotNetworkNode d
-      where
-        isNotNetworkNode i = (<=1) . olength . parentRefs $ refs ! i
-        refs = references ! dag
-
-    rootEdges           = tabulateRootIncidentEdgeset dag
-    ancestoralEdgeSets  = references $ tabulateAncestoralEdgesets dag
-    descendantEdgeSets  = references $ tabulateDescendantEdgesets dag
-    completeEdgeSet     = getEdges dag `difference` rootEdges
-    f k     = foldMapWithKey (g k) . mapWithKey (h k) . childRefs
-    g j k   = foldMap (\x -> S.singleton ((j,k), x))
-    h j k v = possibleEdgeSet j k `difference` v
-    possibleEdgeSet i j = completeEdgeSet `difference` (singletonEdgeSet (i,j) <> singletonEdgeSet (j,i))
-
-
--- |
--- Find all edges adjacent to root nodes.
-tabulateRootIncidentEdgeset :: ReferenceDAG d e n -> EdgeSet (Int,Int)
-tabulateRootIncidentEdgeset dag = foldMap f $ rootRefs dag
-  where
-    f i = foldMap (\e -> singletonEdgeSet (i,e)) kids
-      where
-        kids = IM.keys . childRefs $ references dag ! i
-
-
--- |
--- Gather all paths from a root node to each node in the graph.
-tabulateAncestoralEdgesets :: ReferenceDAG d e n -> ReferenceDAG () (EdgeSet (Int,Int)) ()
-tabulateAncestoralEdgesets dag =
-    RefDAG
-    { references = memo
-    , rootRefs   = rootRefs dag
-    , graphData  = defaultGraphMetadata $ graphData dag
-    }
-  where
-    refs = references dag
-    memo = V.generate (length refs) g
-    g i =
-        IndexData
-        { nodeDecoration = ()
-        , parentRefs     = parentVals
-        , childRefs      = zipWith (<>) childShape (getNetworkEdgeDatum i)
-        }
-      where
-        childShape = ancestorDatum <$ childRefs point
-        point      = refs ! i
-        parentVals = parentRefs point
-        ancestorDatum =
-            case otoList parentVals of
-              []    -> mempty
-              [x]   -> getPreviousDatums x i
-              x:y:_ -> getPreviousDatums x i `union` getPreviousDatums y i
-
-    getPreviousDatums i j = childRefs point ! j <> other
-      where
-        point = memo ! i
-        -- This is the step where new information is added to the accumulator
-        other = singletonEdgeSet (i,j)
-
-    -- We can't let a network edge form a new network edge with it's incident
-    -- network edge. Down that road lies infinite recursion.
-    getNetworkEdgeDatum i = mapWithKey f . childRefs $ refs ! i
-      where
-        f k _ =
-          case otoList . parentRefs $ refs ! k of
-            []  -> mempty
-            [_] -> mempty
-            xs  ->
-              case filter (/=i) xs of
-                []  -> mempty
-                x:_ -> singletonEdgeSet (x,k)
-
-
--- |
--- Gather all paths from a leaf node to each node in the graph.
-tabulateDescendantEdgesets :: ReferenceDAG d e n -> ReferenceDAG () (EdgeSet (Int,Int)) ()
-tabulateDescendantEdgesets dag =
-    RefDAG
-    { references = memo
-    , rootRefs   = rootRefs dag
-    , graphData  = defaultGraphMetadata $ graphData dag
-    }
-  where
-    refs = references dag
-    memo = V.generate (length refs) g
-    g i =
-        IndexData
-        { nodeDecoration = ()
-        , parentRefs     = parentRefs point
-        , childRefs      = descendantDatum <$ childVals
-        }
-      where
-        point     = refs ! i
-        childVals = childRefs point
-        descendantDatum =
-            case IM.keys childVals of
-              []    -> mempty
-              [x]   -> getPreviousDatums i x
-              x:y:_ -> getPreviousDatums i x `union` getPreviousDatums i y
-
-    getPreviousDatums _ j = fold (childRefs point) <> other
-      where
-        point = memo ! j
-        -- This is the step where new information is added to the accumulator
-        other = foldMap (\x -> singletonEdgeSet (j,x)) . IM.keys $ childRefs point
-
-
--- |
 -- Contruct the intermediate 'BinaryRenderingTree' data type for a given 'ReferenceDAG'.
 --
 -- The first parameter is a rendering function for the leaves.
@@ -1418,3 +1301,343 @@ parentsAndChildren i dag = (ps, cs)
     iPoint = references dag ! fromEnum i
     ps = parentRefs iPoint
     cs = childRefs  iPoint
+
+
+-- |
+-- Get the indices of all leaf nodes in the ReferenceDAG.
+leafIndices :: ReferenceDAG d e n -> IntSet
+leafIndices dag = foldMapWithKey leafTest $ dag ^. _references
+  where
+    leafTest :: Int -> IndexData e n -> IntSet
+    leafTest index nodeDatum =
+      if null (childRefs nodeDatum)
+        then IS.singleton index
+        else  mempty
+
+-- |
+-- A function that recursively builds (in a postorder fashion) a generating function
+-- to be consumed as reference data. The function returned uses open recursion
+-- (for memoization purposes), in the form of a 'DVector'.
+dVectorPostorder
+  :: forall a d e n
+   . (  ChildContext a       --  Child values
+     -> (Int, IndexData e n) --  Current index information
+     -> a                    --  Index data
+     )
+  -> ReferenceDAG d e n -> DVector a
+dVectorPostorder indexFn dag = DVector f
+  where
+    refs        = dag ^. _references
+    leafInds    = leafIndices dag
+
+ -- A generate function with open recursion
+    f :: (Int -> a) -> Int -> a
+    f recurseFn ind =
+      if ind `IS.member` leafInds
+        then indexFn NoChildren (ind, refs ! ind)  -- inductive case updating leaves
+        else
+          case otoChildContext . IM.keysSet $ (refs ! ind) ^. _childRefs of
+            NoChildren
+              -> error "Non-leaf node without children!"
+            OneChild childInd                           -- recursively apply the function
+              -> indexFn
+                   (OneChild (recurseFn childInd))
+                   (ind, refs ! ind)
+            TwoChildren childInd1 childInd2                 -- Same as above.
+              -> indexFn
+                   (TwoChildren (recurseFn childInd1) (recurseFn childInd2))
+                   (ind, refs ! ind)
+
+
+-- |
+-- A function that recursively builds (in a preorder fashion) a generating function
+-- to be consumed as reference data. The function returned uses open recursion
+-- (for memoization purposes), in the form of a 'DVector'.
+dVectorPreorder
+  :: forall a d e n
+   . (  ParentContext a      --  Parent values
+     -> (Int, IndexData e n) --  Current index information
+     -> a                    --  Index data
+     )
+  -> ReferenceDAG d e n -> DVector a
+dVectorPreorder indexFn dag = DVector f
+  where
+    refs        = dag ^. _references
+    rootInds    = rootRefs dag
+
+ -- A generate function with open recursion
+    f :: (Int -> a) -> Int -> a
+    f recurseFn ind =
+      if ind `elem` rootInds
+        then indexFn NoParent (ind, refs ! ind)  -- base case updating roots
+        else
+          case otoParentContext $ (refs ! ind) ^. _parentRefs of
+            NoParent
+              -> error "Non-root node without parents!"
+            OneParent parInd                           -- recursively apply the function
+              -> indexFn
+                   (OneParent (recurseFn parInd))
+                   (ind, refs ! ind)
+            TwoParents parInd1 parInd2                 -- Same as above.
+              -> indexFn
+                   (TwoParents (recurseFn parInd1) (recurseFn parInd2))
+                   (ind, refs ! ind)
+
+
+-- |
+-- This computes, in a nodal context, the set of ancestral
+-- edges from the parent edge sets and current node data.
+ancestralEdgeSetContextFn
+  :: ParentContext (EdgeSet (Int, Int))  -- ^ Parent ancestral edge sets
+  -> (Int, IndexData e n)                -- ^ Current node data
+  -> EdgeSet (Int, Int)                  -- ^ Current node edge sets
+ancestralEdgeSetContextFn ancestralEdgeSets (currInd, nodeDatum) =
+    case ancestralEdgeSets of
+      NoParent                   -> mempty
+      OneParent ancestralEdgeSet -> currentEdgeSet <> ancestralEdgeSet
+
+      TwoParents ancestralEdgeSet1 ancestralEdgeSet2
+        -> currentEdgeSet <> ancestralEdgeSet1 <> ancestralEdgeSet2
+  where
+    parRefs        = nodeDatum ^. _parentRefs
+    currentEdgeSet = makeParentEdgeSet currInd parRefs
+
+
+
+
+-- |
+-- This computes, in a nodal context, the set of descendent
+-- network nodes from the child descendent network sets and current node data.
+-- It does this via a 'traversal with state' which keeps track of whether a child
+-- node has only a single child and so is a descendant network node.
+descendantNetworkNodesContextFn
+  ::  ChildContext (IntSet, Maybe Int)   -- ^ child descendent network set and maybe node
+  -> (Int, IndexData e n)                -- ^ Current node data
+  -> (IntSet, Maybe Int)                 -- ^ Current descendant edge sets and maybe node
+descendantNetworkNodesContextFn descendantNetworkNodes (currInd, _) =
+  case descendantNetworkNodes of
+    NoChildren                              -> (mempty, Nothing)
+ -- If a node has a single child then it is a network node and so is added
+ -- to the set of nodes to be included in the parent descendant sets.
+    OneChild (networkNodes, optNode)
+      -> case optNode of
+          Nothing -> (networkNodes, Just currInd)
+       -- This case should never happen as it corresponds to a
+       -- current network node with descendant network node.
+          Just n  -> (IS.singleton n <> networkNodes, Just currInd)
+    TwoChildren
+      (networkNodes1, optNode1)  (networkNodes2, optNode2)
+      -> case (optNode1, optNode2) of
+           (Nothing , Nothing)
+             -> (networkNodes1 <> networkNodes2, Nothing)
+           (Just n1 , Nothing)
+             -> (IS.singleton n1 <> networkNodes1 <> networkNodes2, Nothing)
+           (Nothing , Just n2)
+             -> (IS.singleton n2 <> networkNodes1 <> networkNodes2, Nothing)
+           (Just n1 , Just n2)
+             -> (     IS.singleton n1
+                   <> IS.singleton n2
+                   <> networkNodes1
+                   <> networkNodes2
+                , Nothing
+                )
+
+
+-- |
+-- This computes, in a nodal context, the set of ancestral
+-- nodes which are incident to a root node set. It does this
+-- via a 'traversal with state' passing a boolean value of
+-- whether a node is a root node.
+ancestralRootIncidentNodesContextFn
+  :: ParentContext (IntSet,Bool)  -- ^ parent root node set and state information
+  -> (Int, IndexData e n)         -- ^ Current node data
+  -> (IntSet, Bool)               -- ^ Current node edge sets
+ancestralRootIncidentNodesContextFn ancestralRootNodes (currInd, nodeDatum) =
+  case ancestralRootNodes of
+    NoParent
+      -> (mempty, True)
+    OneParent (parAncestralSet, incidentToRoot)
+      -> if incidentToRoot
+           then (IS.singleton currInd, False)
+           else (parAncestralSet, False)
+    TwoParents
+      (parAncestralSet1, incidentToRoot1)
+      (parAncestralSet2, incidentToRoot2)
+        -> if incidentToRoot1 && incidentToRoot2
+             then (IS.singleton currInd <> parAncestralSet1 <> parAncestralSet2, False)
+             else (parAncestralSet1 <> parAncestralSet2, False)
+
+-- |
+-- Generate a vector of graph data for finding the candidate network edges in a memoized
+-- fashion.
+tabulateNetworkInformation
+  :: ReferenceDAG d e n
+  -> Vector
+        ( EdgeSet (Int, Int)    -- Ancestral Edge set
+        , (IntSet, Maybe Int)   -- Pair of Descendant Network Node Information and
+                                -- potential network node index
+        , (IntSet, Bool)        -- Ancestral root incident nodes and root node bool
+        )
+tabulateNetworkInformation dag =
+  let
+    dVectorAncestralEdge         = dVectorPreorder  ancestralEdgeSetContextFn       dag
+    dVectorDescendantNet         = dVectorPostorder descendantNetworkNodesContextFn dag
+    dVectorAncestralRootIncident = dVectorPreorder  ancestralRootIncidentNodesContextFn dag
+
+    dVectorNetInfo
+      = DV.zip3
+          dVectorAncestralEdge
+          dVectorDescendantNet
+          dVectorAncestralRootIncident
+  in
+    generateMemo lengthRefs dVectorNetInfo
+  where
+    lengthRefs = length $ dag ^. _references
+
+
+
+-- |
+-- Find all candidate network edges in a DAG.
+candidateNetworkEdges :: ReferenceDAG d e n -> Set ((Int, Int), (Int,Int))
+candidateNetworkEdges dag = S.fromList candidateEdgesList
+  where
+    completeEdges      = toList $ getEdges dag
+    rootIndices        = IS.fromList . toList . rootRefs $ dag
+    networkNodes       = gatherDescendantNetworkNodes rootIndices networkInformation
+
+  -- This vector contains all the information needed for the various edge
+ -- compatibility criteria.
+    networkInformation = tabulateNetworkInformation dag
+
+
+ -- Gets all pairs of distinct edges from the edge set that can be
+ -- compatably added to the network.
+    candidateEdgesList :: [((Int, Int), (Int, Int))]
+    candidateEdgesList =
+        do
+       -- collect distinct edge pairs.
+          (e1 : es)       <- tails completeEdges
+          e2@(src2, tgt2) <- es
+          guard (symmetricCompatibility e1 e2)
+
+          let  e1e2Bool = not (hasIncidentNetworkNode e2) && posetalCompatibility e1 e2
+          let  e2e1Bool = not (hasIncidentNetworkNode e1) && posetalCompatibility e2 e1
+          case (e1e2Bool, e2e1Bool) of
+            (True, True)   -> pure (e1,e2) <|> pure (e2,e1)
+            (True, False)  -> pure (e1, e2)
+            (False, True)  -> pure (e2, e1)
+            (False, False) -> Alt.empty
+
+ -- Checks historical compatability in a symmetrical fashion.
+    symmetricCompatibility :: (Int, Int) -> (Int, Int) -> Bool
+    symmetricCompatibility e1@(src1, tgt1) e2@(src2, tgt2) =
+      let
+     -- Ancestral/descendant network information:
+        e1AncestralEdges = proj3_1 $ networkInformation ! tgt1
+        e2AncestralEdges = proj3_1 $ networkInformation ! tgt2
+
+        e1NoRootSource
+          = not $ src1 `IS.member` rootIndices
+        e2NoRootSource
+          = not $ src2 `IS.member` rootIndices
+      in
+          -- First check if the two edges are from the same parent to short circuit
+          -- faster in this case.
+             src1 /= src2
+          -- or if either edge is ancestral to the other.
+          && (e1AncestralEdges `disjoint` e2AncestralEdges)
+          && e1NoRootSource && e2NoRootSource
+
+ -- This checks posetal/historical compability for inserting an edge from
+ -- e1 to e2.
+    posetalCompatibility :: (Int, Int) -> (Int, Int) -> Bool
+    posetalCompatibility e1@(src1, tgt1) e2@(src2, tgt2) =
+ -- We check historical compatibility by checking if the
+ -- descendant network node of tgt1 is "historically compatible" to the descendant
+ -- network nodes of src2. This is because these are the historical
+ -- coniditons the new nodes will inherit:
+ --                o  ── src1     ┌─────────────────────────────────────────────────┐
+ --               /             ┌─│ New parent network node of newTgt has the same  │
+ --              /              │ │ descendant network events as tgt1.              │
+ --             o  ── newSrc ───┘ └─────────────────────────────────────────────────┘
+ --            /  \______________________
+ --           /                          \
+ --          o ── tgt1                    \
+ --                                     [...]
+ --                                         \              ┌───────── o ── src2
+ --                                          \____________ │ __      /
+ --    ┌────────────────────────────────────────────┐      │   \    /
+ --    │  src2 is the other new parent network node │      │    \  /
+ --    │  of newTgt and so must be potentially      │──────┘      o  ── newTgt
+ --    │  historically coincident with newSrc.      │            /
+ --    └────────────────────────────────────────────┘           /
+ --                                                            o  ── tgt2
+ --   ┌─────────────────────────────────────┐
+ --   │ Potentially historically coincident │
+ --   └─────────────────────────────────────┘
+ --   This  means that any descendant network node of src2 cannot be a descendant
+ --   network node to a non-root node ancestral to src1 as this would lead to
+ --   non-transitivty in the implied oredering. Similarly we cannot have that a
+ --   descendant network node of tgt1 (which has the same  descendant nodes as
+ --   newSrc other than newTgt) equal to a descendant network node of a non-root
+ --   node ancestral to src2.
+      let
+     -- Ancestral/descendant network information:
+        e1AncestralRootIncidentNodes
+          = (fst . proj3_3 $ networkInformation ! tgt1)
+        e2AncestralRootIncidentNodes
+          = (fst . proj3_3 $ networkInformation ! tgt2)
+
+        e1SrcAncestralNetworkNodes
+          = gatherDescendantNetworkNodes e1AncestralRootIncidentNodes networkInformation
+        e2SrcAncestralNetworkNodes
+          = gatherDescendantNetworkNodes e2AncestralRootIncidentNodes networkInformation
+
+        e1TgtDescendantNetworkNodes
+          = fst . proj3_2 $ networkInformation ! tgt1
+        e2SrcDescendantNetworkNodes
+          = fst . proj3_2 $ networkInformation ! src2
+
+     -- Boolean tests:
+      {-- See TODO below
+        e1HasRootSource
+          = src1 `IS.member` rootIndices
+        e2HasRootSource
+          = src2 `IS.member` rootIndices --}
+        e1AncestralTest
+          = e2SrcDescendantNetworkNodes `IS.disjoint` e1SrcAncestralNetworkNodes
+        e2AncestralTest
+          = e1TgtDescendantNetworkNodes `IS.disjoint` e2SrcAncestralNetworkNodes
+      in
+      -- This is the correct version if we allow root node as edge source
+      --
+      -- TODO: Revisit this network edge constraint.
+      -- We will re-evaluate how to add network edges incident to the root after
+      -- some of the core data structures related to the topology and character
+      -- matrix
+      {-
+        case (e1HasRootSource, e2HasRootSource) of
+          (True , True ) -> False
+          (True , False) -> e2AncestralTest
+          (False, True ) -> e1AncestralTest
+          (False, False) -> e1AncestralTest && e2AncestralTest
+      -}
+        e1AncestralTest && e2AncestralTest
+
+
+    hasIncidentNetworkNode :: (Int, Int) -> Bool
+    hasIncidentNetworkNode e@(src,tgt) =
+         tgt `IS.member` networkNodes
+      || src `IS.member` networkNodes
+
+
+
+
+-- |
+-- Helper function to get all descendent network nodes from an `IntSet` of nodes.
+gatherDescendantNetworkNodes
+  :: IntSet                     -- ^ Node set
+  -> Vector (a, (IntSet, b), c) -- ^ Vector tuple with network node indices
+  -> IntSet                     -- ^ All descendant network nodes
+gatherDescendantNetworkNodes inds vect
+  = ofoldMap (\ind -> fst . proj3_2 $ vect ! ind) inds