+++ test/TestSuite/LibraryTests.hs
@@ -5,6 +5,7 @@ module TestSuite.LibraryTests
 import qualified Analysis.Parsimony.Dynamic.DirectOptimization.Pairwise.Test as Pairwise
 import qualified Bio.Character.Encodable.Dynamic.Test                        as DynamicChar
 import qualified Bio.Character.Encodable.Static.Test                         as StaticChar
+import qualified Bio.Graph.ReferenceDAG.Test                                 as ReferenceDAG
 import qualified Control.Evaluation.Test                                     as Evaluation
 import qualified Data.Alphabet.Test                                          as Alphabet
 import qualified Data.TCM.Test                                               as TCM
@@ -19,4 +20,5 @@ testSuite = testGroup "Library Test Suite"
     , DynamicChar.testSuite
     , StaticChar.testSuite
     , Pairwise.testSuite
+    , ReferenceDAG.testSuite
     ]