+++ test/TestSuite/ScriptTests.hs
@@ -1,31 +1,35 @@
 {-# LANGUAGE FlexibleContexts  #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE Strict            #-}
 {-# LANGUAGE TypeFamilies      #-}
 
 module TestSuite.ScriptTests
   ( testSuite
   ) where
 
-import Control.Arrow              ((&&&))
 import Control.DeepSeq
+import Data.Bifunctor             (first)
 import Data.Char                  (isSpace)
 import Data.Either
-import Data.Foldable
-import Data.Functor               (($>))
+import Data.Functor               (void, ($>))
 import Data.Scientific            hiding (scientific)
 import Data.Text                  (Text)
+import Data.Text.IO               (readFile)
 import Data.Void                  (Void)
 import Numeric.Extended.Real
+import Prelude                    hiding (readFile, writeFile)
+import System.Exit
+import System.Process
 import Test.Tasty
 import Test.Tasty.HUnit
+import TestSuite.SubProcess
 import Text.Megaparsec
 import Text.Megaparsec.Char
 import Text.Megaparsec.Char.Lexer (scientific)
-import Turtle                     hiding (char, many, parallel, satisfy, wait, x)
 
 
-testSuite :: IO TestTree
-testSuite = testGroup "Script Test Suite" <$> sequenceA
+testSuite :: TestTree
+testSuite = testGroup "Script Test Suite"
   [ scriptCheckCost 50.46
         "datasets/continuous/single-block/arthropods.pcg"
         "datasets/continuous/single-block/cost.data"
@@ -152,7 +156,6 @@ testSuite = testGroup "Script Test Suite" <$> sequenceA
   , scriptCheckCost 11036
       "datasets/dynamic/single-block/protein/L1-norm/invertebrates.pcg"
       "datasets/dynamic/single-block/protein/L1-norm/cost.data"
-{--
   , scriptCheckCost 1132
       "datasets/dynamic/single-block/protein/discrete/invertebrates.pcg"
       "datasets/dynamic/single-block/protein/discrete/cost.data"
@@ -162,11 +165,9 @@ testSuite = testGroup "Script Test Suite" <$> sequenceA
   , scriptCheckCost 1241
         "datasets/dynamic/single-block/protein/2-1/invertebrates.pcg"
         "datasets/dynamic/single-block/protein/2-1/cost.data"
---}
   , scriptCheckCost 3413
       "datasets/dynamic/single-block/slashes/L1-norm/test.pcg"
       "datasets/dynamic/single-block/slashes/L1-norm/cost.data"
-{--
   , scriptCheckCost 197
         "datasets/dynamic/single-block/slashes/discrete/test.pcg"
         "datasets/dynamic/single-block/slashes/discrete/cost.data"
@@ -182,22 +183,22 @@ testSuite = testGroup "Script Test Suite" <$> sequenceA
   , scriptCheckCost 488
         "datasets/dynamic/single-block/slashes/levenshtein/test.pcg"
         "datasets/dynamic/single-block/slashes/levenshtein/cost.data"
-  , scriptCheckCost 197
+  , scriptCheckCost 133
         "datasets/dynamic/single-block/large-mix/discrete/test.pcg"
         "datasets/dynamic/single-block/large-mix/discrete/cost.data"
-  , scriptCheckCost 2042
+  , scriptCheckCost 7185
         "datasets/dynamic/single-block/large-mix/L1-norm/test.pcg"
         "datasets/dynamic/single-block/large-mix/L1-norm/cost.data"
-  , scriptCheckCost 254
+  , scriptCheckCost 164
         "datasets/dynamic/single-block/large-mix/1-2/test.pcg"
         "datasets/dynamic/single-block/large-mix/1-2/cost.data"
-  , scriptCheckCost 228
+  , scriptCheckCost 172
         "datasets/dynamic/single-block/large-mix/2-1/test.pcg"
         "datasets/dynamic/single-block/large-mix/2-1/cost.data"
-  , scriptCheckCost 671
+  , scriptCheckCost 367
         "datasets/dynamic/single-block/large-mix/hamming/test.pcg"
         "datasets/dynamic/single-block/large-mix/hamming/cost.data"
-  , scriptCheckCost 488
+  , scriptCheckCost 213
         "datasets/dynamic/single-block/large-mix/levenshtein/test.pcg"
         "datasets/dynamic/single-block/large-mix/levenshtein/cost.data"
   , scriptCheckCost 197
@@ -218,7 +219,6 @@ testSuite = testGroup "Script Test Suite" <$> sequenceA
   , scriptCheckCost 488
         "datasets/dynamic/single-block/huge-mix/levenshtein/test.pcg"
         "datasets/dynamic/single-block/huge-mix/levenshtein/cost.data"
---}
   , scriptFailure "datasets/unmatched-leaf-taxon/test.pcg"
   , scriptFailure "datasets/unmatched-tree-taxon/test.pcg"
   , scriptFailure "datasets/duplicate-leaf-taxon/test.pcg"
@@ -230,37 +230,32 @@ testSuite = testGroup "Script Test Suite" <$> sequenceA
   ]
 
 
-scriptTest
-  :: String                          -- ^ Script File
-  -> [String]                        -- ^ Expected Output Files
-  -> (Either Int [Text] -> TestTree) -- ^ Build a TestTree from the resulting
-                                     -- output file contents or ExitStatus code
-  -> IO TestTree
-scriptTest scriptPath outputPaths testLogic = testLogic <$> runExecutable scriptPath outputPaths
-
-
 scriptCheckCost
-  :: ExtendedReal                    -- ^ Expected cost ∈ [0, ∞]
-  -> String                          -- ^ Script File
-  -> String                          -- ^ Expected output file containing the cost
-  -> IO TestTree
-scriptCheckCost expectedCost scriptPath outputPath = scriptTest scriptPath [outputPath] $ testCase scriptPath . checkResult
-  where
-    checkResult (Left     exitCode) = assertFailure $ "Script failed with exit code: " <> show exitCode
-    checkResult (Right          []) = assertFailure "No files were returned despite supplying one path!"
-    checkResult (Right (outFile:_)) =
-        case force $ parseCost outFile of
-          Nothing   -> assertFailure "No cost found in the output file!"
-          Just cost -> cost @?= expectedCost
+  :: ExtendedReal -- ^ Expected cost ∈ [0, ∞]
+  -> String       -- ^ Script File
+  -> String       -- ^ Expected output file containing the cost
+  -> TestTree
+scriptCheckCost expectedCost scriptPath outputPath = testCase scriptPath $ do
+    v <- runExecutable scriptPath [outputPath]
+    case v of
+      Left     exitCode -> assertFailure $ "Script failed with exit code: " <> show exitCode
+      Right          [] -> assertFailure "No files were returned despite supplying one path!"
+      Right (outData:_) -> case force $ parseCost outputPath outData of
+                             Left  pErr -> assertFailure $ "No cost found in the output file!\n" <> pErr <> show outData
+                             Right cost -> cost @?= expectedCost
 
 
-parseCost :: Text -> Maybe ExtendedReal
-parseCost = parseMaybe fileSpec
+parseCost
+  :: FilePath -- ^ The path of the input file
+  -> Text     -- ^ The text stream to parse
+  -> Either String ExtendedReal
+parseCost path str = first (parseErrorPretty' str) parseResult
   where
+    parseResult :: Either (ParseError Char Void) ExtendedReal
+    parseResult = parse fileSpec path $ force str
+
     fileSpec =  many (try (ignoredLine <* notFollowedBy costLine))
-             *> ignoredLine
-             *> costLine
-             <* many ignoredLine
+             *> (try costLine <|> (ignoredLine *> costLine))
 
     costLine :: MonadParsec Void Text m => m ExtendedReal
     costLine =  many inlineSpace
@@ -269,12 +264,11 @@ parseCost = parseMaybe fileSpec
              *> extendedReal
              <* ignoredLine
 
-
     extendedReal :: MonadParsec Void Text m => m ExtendedReal
     extendedReal = (fromFinite . toRealFloat <$> scientific) <|> (char '∞' $> infinity)
 
     ignoredLine  :: MonadParsec Void Text m => m ()
-    ignoredLine  = many inlineChar *> newlineChar
+    ignoredLine  = many inlineChar *> newlineChar -- (newlineChar <|> eof)
 
     inlineSpace  :: MonadParsec Void Text m => m ()
     inlineSpace  = void $ satisfy (\x -> isSpace x && x /= '\n')
@@ -286,9 +280,10 @@ parseCost = parseMaybe fileSpec
     newlineChar  = void $ char '\n'
 
 
-
-scriptFailure :: String -> IO TestTree
-scriptFailure scriptPath = scriptTest scriptPath [] (testCase scriptPath . assertBool "Expected script failure" . isLeft)
+scriptFailure :: String -> TestTree
+scriptFailure scriptPath = testCase scriptPath $ do
+    v <- runExecutable scriptPath []
+    assertBool "Expected script failure" $ isLeft v
 
 
 -- |
@@ -302,18 +297,106 @@ runExecutable
   -> IO (Either Int [Text]) -- ^ Resulting file contents of the specified output
                             --   files, or the exit code if the script failed
 runExecutable scriptStr outputPaths = do
-    startingDirectory <- pwd
-    cd scriptDirectory
-    (exitCode, _) <- shellStrict ("stack exec pcg -- --input " <> scriptText <> " --output test.log") mempty
-    cd startingDirectory
+--
+{-
+    dbFile <- openFile "output.debug" WriteMode
+    hSetBuffering dbFile NoBuffering
+    hPutStrLn dbFile startingDirectory
+    hPutStrLn dbFile scriptDirectory
+    hPutStrLn dbFile absScriptDirectory
+    _      <- hClose dbFile
+-}
+    ctx <- constructProcess scriptStr
+    (exitCode, _, _) <- readCreateProcessWithExitCode (process ctx) mempty
+--                            (defineProcess absScriptDirectory command) mempty
+--                            (defineProcess scriptDirectory command) mempty
+    _ <- destructProcess ctx
     case exitCode of
-      ExitFailure v -> pure $ Left v
-      _             -> Right <$> traverse (readTextFile . decodeString) outputPaths
+      ExitFailure v -> pure  $ Left v
+      _             -> force . Right <$> traverse readFile outputPaths
+--  where
+--    command :: String
+--    command = "stack exec pcg -- --input " <> scriptFile <> " --output test.log"
+
+
+{-
+outLogFileName :: String
+outLogFileName = "log.out"
+
+
+errLogFileName :: String
+errLogFileName = "log.err"
+
+
+data  ScriptContext
+    = ScriptContext
+    { process :: CreateProcess
+    , outPath :: FilePath
+    , errPath :: FilePath
+    }
+
+
+constructProcess
+  :: FilePath -- ^ Relative path to the PCG script
+  -> IO ScriptContext
+constructProcess scriptStr = do
+    prefix <- makeAbsolute scriptDirectory
+    let outFilePath = prefix </> outLogFileName
+    let errFilePath = prefix </> errLogFileName
+    let commandStr  = unwords
+                    [ "stack exec pcg -- --input"
+                    , scriptFileName
+                    , "--output"
+                    , outFilePath
+                    , "2>"
+                    , errFilePath
+                    ]
+
+    let p = CreateProcess
+            { cmdspec            = ShellCommand commandStr
+            , cwd                = Just scriptDirectory
+            , env                = Nothing
+            -- Do not use the stream handles, they do not work with Tasty
+            , std_in             = NoStream
+            , std_out            = NoStream
+            , std_err            = NoStream
+            , close_fds          = True
+            , create_group       = False
+            , delegate_ctlc      = False
+            , detach_console     = False
+            , create_new_console = False
+            , new_session        = False
+            , child_group        = Nothing
+            , child_user         = Nothing
+            , use_process_jobs   = False
+            }
+
+    pure ScriptContext
+        { process = p
+        , outPath = outFilePath
+        , errPath = errFilePath
+        }
   where
-    scriptText = either id id $ toText scriptFile
+    (scriptDirectory, scriptFileName) = breakScriptPath scriptStr
 
-    (scriptDirectory, scriptFile) = breakScriptPath $ decodeString scriptStr
-
-    breakScriptPath = (collapse . foldl' (</>) defaultDirectory . init &&& last) . splitDirectories
+    breakScriptPath = (normalise . foldl' (</>) defaultDirectory . init &&& last) . splitDirectories
       where
-        defaultDirectory = decodeString "."
+        defaultDirectory = "."
+
+
+destructProcess :: ScriptContext -> IO ()
+destructProcess ctx = mapM_ cleanUpHandle $
+    [ (std_out . process &&& outPath)
+    , (std_err . process &&& errPath)
+    ] <*> [ ctx ]
+  where
+    cleanUpHandle (s, p) = do
+        -- If we have an opened handle, close it
+        _ <- case s of
+               UseHandle h -> hClose h
+               _           -> pure ()
+        n <- getFileSize p
+        if n == 0 -- If no data was written, delete the file
+        then removeFile p
+        else pure ()
+-}